BigQuery-Python/requirements.txt
#pip install -r requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.11.0
xmlschema==2.4.0
